error register

+++ src/main/java/application/api/web/NewApiController.java
@@ -23,7 +23,7 @@ public class NewApiController {
     }
 
     @DeleteMapping("/api/new")
-    public void deleteNew(@RequestBody long[] ids) {
+    public void deleteNew(@RequestBody int[] ids) {
         newService.delete(ids);
     }
 }
