+++ src/main/java/application/repository/CategoryRepository.java
@@ -3,6 +3,6 @@ package application.repository;
 import application.entity.CategoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<CategoryEntity,Long> {
+public interface CategoryRepository extends JpaRepository<CategoryEntity,Integer> {
     CategoryEntity findOneByCode(String code);
 }
