+++ src/main/java/application/repository/NewRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NewRepository extends JpaRepository<NewEntity,Long> {
+public interface NewRepository extends JpaRepository<NewEntity,Integer> {
 }
