+++ src/main/java/application/service/INewService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface INewService {
     List<NewDTO> findAll();
-    NewDTO findById(long id);
+    NewDTO findById(int id);
     NewDTO save(NewDTO newDTO);
-    void delete(long[] ids);
+    void delete(int[] ids);
 }
